refactor: переместил служебные методы из класса Director в DirectorController и DirectorDbService
